@@ -4,11 +4,11 @@
   "private": true,
   "dependencies": {
     "clarifai": "^2.9.0",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
-    "react-particles-js": "^2.3.0",
-    "react-scripts": "2.1.0",
-    "react-tilt": "^0.1.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-particles": "^2.12.2",
+    "react-scripts": "5.0.1",
+    "react-tilt": "^1.0.2",
     "tachyons": "^4.11.1"
   },
   "scripts": {
