@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Particles from 'react-particles-js';
+import Particles from 'react-particles';
 import 'tachyons';
 import './App.css';
 import Navigation from './components/Navigation/Navigation.js';
